+++ projects/ngx-uc/src/public-api.ts
@@ -4,6 +4,7 @@
 
 export * from './lib/ngx-uc-zoom-view/uc-zoom-view.component';
 export * from './lib/ngx-uc-zoom-view/uc-zoom-view-config';
+export * from './lib/ngx-uc-zoom-view/uc-zoom-view-events';
 export * from './lib/uc-css-classes';
 export * from './lib/uc-coordinates';
 export * from './lib/ngx-uc.module';